+++ CHANGES.md
@@ -32,6 +32,10 @@
     )
     ```
 
+### titiler.xarray
+
+* change `get_variable.drop_dim` parameter type from `str` to `List[str]` **breaking change**
+
 ## 0.21.1 (2025-01-29)
 
 ### titiler.core